requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
blurhash==1.1.4
certifi==2022.9.24
charset-normalizer==2.1.1
datasets==2.7.0
decorator==5.1.1
dill==0.3.6
evaluate==0.3.0
filelock==3.8.0
frozenlist==1.3.3
fsspec==2022.11.0
huggingface-hub==0.11.0
idna==3.4
Mastodon.py==1.8.0
multidict==6.0.2
multiprocess==0.70.14
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyarrow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic==0.4.27
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
responses==0.18.0
sentencepiece==0.1.97
six==1.16.0
tokenizers==0.13.2
-e git+https://github.com/pytorch/pytorch.git@7c98e70d44abc7a1aead68b6ea6c8adc8c554db5#egg=torch
-e git+https://github.com/pytorch/vision.git@5ce4506ac43c8b1dc1736ed9e51c58e0e29f5237#egg=torchvision
tqdm==4.64.1
transformers==4.23.1
twitter==1.19.6
typing-extensions==4.4.0
urllib3==1.26.12
xxhash==3.1.0
yarl==1.8.1
